Threat translation: write threat as int64, not uint32
Modern SMSG_THREAT_UPDATE / HIGHEST_THREAT_UPDATE pack the threat field as
int64 (8 bytes), not uint32. Verified against TrinityCore master and the
Frostshake/TrinityCoreClassic 1.14.0.40618 fork — both write
  _worldPacket << threat;
where the type is int64.

Symptom of the bug: testing with hunter + pet, Tiny Threat displayed pet
threat as ~7.1 billion instead of ~2.1k. The proxy's internal raw threat
totals (in threat.damage_added diagnostics) showed correct ~2154 — so the
inflation was on the wire, not in the calculation.

Mechanism: the writer emitted 4 bytes; the client read 8. The high 4
bytes of each int64 came from the next entry's GUID-mask bytes,
producing arbitrarily large garbage when the threat list had >1 entry.
The Phase 1 smoke test happened to display correctly only because there
was a single threater and the high bytes were zero-padded buffer.

Fix: change ThreatInfo.Threat from uint to long, write via
_worldPacket.WriteInt64. Update ToWireThreat clamp ceiling and the smoke
test literals (1234500u -> 1234500L) to match.

+++ HermesProxy/World/Server/Packets/CombatPackets.cs
@@ -348,13 +348,24 @@ class PartyKillLog : ServerPacket, ISpanWritable
 // event) populate, which in turn drives addons like Details TinyThreat and
 // Threat Plates without any addon-side modification.
 //
-// Wire format follows TrinityCore/CypherCore convention. Note that the modern
-// protocol packs threat values × 100 (Classic Era addons divide by 100 on read),
-// so emitters must scale up before writing.
+// Wire format follows TrinityCore Classic 1.14 convention:
+//   PackedGuid128 unitGUID
+//   uint32 count
+//   foreach threater:
+//     PackedGuid128 threaterGUID
+//     int64 threat  <-- 8 BYTES, not 4. Modern protocol uses int64 even though
+//                       practical threat values fit in 32 bits; reading 4 here
+//                       causes the client to gobble the next entry's GUID-mask
+//                       bytes into the high dword and produce billions-scale
+//                       garbage values. Verified against TrinityCore master
+//                       and Frostshake/TrinityCoreClassic 1.14.0.40618.
+//
+// Note that the modern protocol packs threat values × 100 (Classic Era addons
+// divide by 100 on read), so emitters must scale up before writing.
 public class ThreatInfo
 {
     public WowGuid128 ThreaterGUID;
-    public uint Threat; // raw threat × 100
+    public long Threat; // raw threat × 100, on-wire int64
 }
 
 public class ThreatUpdatePkt : ServerPacket
@@ -368,7 +379,7 @@ public class ThreatUpdatePkt : ServerPacket
         foreach (var info in ThreatList)
         {
             _worldPacket.WritePackedGuid128(info.ThreaterGUID);
-            _worldPacket.WriteUInt32(info.Threat);
+            _worldPacket.WriteInt64(info.Threat);
         }
     }
 
@@ -388,7 +399,7 @@ public class HighestThreatUpdatePkt : ServerPacket
         foreach (var info in ThreatList)
         {
             _worldPacket.WritePackedGuid128(info.ThreaterGUID);
-            _worldPacket.WriteUInt32(info.Threat);
+            _worldPacket.WriteInt64(info.Threat);
         }
     }
